+++ tests/server.js
@@ -1,3 +1,3 @@
 var Tests = require( './tests.js' );
 
-Tests.Suite( require( '../validator.js' ), require( 'expect.js' ) );
+Tests.Suite( require( '../dist/validator.js' ), require( 'expect.js' ) );